+++ packages/@glimmer/runtime/lib/vm/render-result.ts
@@ -2,7 +2,7 @@ import { Option, LinkedList } from '@glimmer/util';
 import Environment, { Program } from '../environment';
 import { DestroyableBounds, clear } from '../bounds';
 import UpdatingVM, { ExceptionHandler } from './update';
-import { UpdatingOpcode } from '../opcodes';
+import { UpdatingOpcode } from '../updating-opcodes';
 import { Simple } from '@glimmer/interfaces';
 
 export default class RenderResult implements DestroyableBounds, ExceptionHandler {
